Health Check Moved to Socket Thread
In order for the Health CHeck to be more meaningful I removed it's
thread and added it to the same thread as Socket.

In order to make it possible for Health Check Connections and UDP
massages to both be recieved on the same thread and not create
a problem when listening while also not wasting CPU I made the
Socket non blocking and added a select call to let the server know
when there is a new connection or data to read.

Moved the socket error handling code into it's own file.

Added the Socket helper functions to the Socket implementations.